% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lol.R
\name{lol}
\alias{lol}
\title{lol engine}
\description{
Create a \code{lol} object to pass to \link{lolcat} and friends
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{lol$new()}}
\item \href{#method-render}{\code{lol$render()}}
\item \href{#method-render_line}{\code{lol$render_line()}}
\item \href{#method-set_as_default}{\code{lol$set_as_default()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new lol engine
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{lol$new(rate = 0.02, angle = pi/4, aspect = 3, palette = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{rate}}{The rate of colour progression in terms of fraction
of the colour palette per horizontal line when \code{angle = 0}.
Increasing values give increasing rate of colour change.}

\item{\code{angle}}{The angle of the colour change. 0 is horizontal
lines only, \code{pi / 4} (the default) runs the gradient top-left
to bottom-right, \code{pi / 2} changes across columns and so on.
Any numeric value is allowed.}

\item{\code{aspect}}{The assumed aspect ratio of rows to columns (used
in the treatment of theta).}

\item{\code{palette}}{The colour palette to use. This should be a vector
of hex colours that "wrap". We will interpolate along this
vector to select a colour for the output. The default is a
rainbow set of colours.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-render"></a>}}
\if{latex}{\out{\hypertarget{method-render}{}}}
\subsection{Method \code{render()}}{
Render text in colour
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{lol$render(text, reset = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{text}}{Character vector to render}

\item{\code{reset}}{Reset the column on return (i.e, assume text is followed
by a newline)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-render_line"></a>}}
\if{latex}{\out{\hypertarget{method-render_line}{}}}
\subsection{Method \code{render_line()}}{
Render a single line (i.e., single string) in colour
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{lol$render_line(text, reset = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{text}}{Single string to render}

\item{\code{reset}}{Reset the column on return (i.e, assume text is followed
by a newline)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_as_default"></a>}}
\if{latex}{\out{\hypertarget{method-set_as_default}{}}}
\subsection{Method \code{set_as_default()}}{
Set this object as the package default
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{lol$set_as_default()}\if{html}{\out{</div>}}
}

}
}
