% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knitr.R
\name{lol_hooks}
\alias{lol_hooks}
\title{Hook for knitr}
\usage{
lol_hooks(background = "#222222", plain = "#eeeeee")
}
\arguments{
\item{background}{Colour to use for the background (any valid CSS
specification). The default is quite dark grey, which works well
with the default colour theme.}

\item{plain}{Colour to use for unstyled text; practically this
is any knitr pre-output prompt like \verb{#>}. Should contrast
against \code{background}. The default is very light grey.}
}
\value{
Invisibly, an argument-less function that will restore the
previous state before running this function.
}
\description{
Hook for use with knitr. Must have knitr and fansi installed to
use. Must be within a block that has \code{results = "asis"} (though
may be \verb{echo = FALSE)} as we output some style.
}
