% Generated by roxygen2 (4.0.1): do not edit by hand
\name{lolcat}
\alias{lolcat}
\title{Taste The Rainbow}
\usage{
lolcat(..., file = "")
}
\arguments{
\item{...}{Arguments as for \code{\link{cat}}, including
\code{sep}, etc.}

\item{file}{Used to check that we don't put silly output into
actual files.  Base \code{cat} will be used when \code{file} is
given.}
}
\description{
Replacement for cat
}
\examples{
for (i in 1:20) {
  lolcat("hello world\\n")
}
}
\author{
Rich FitzJohn
}

