% Generated by roxygen2 (4.0.1): do not edit by hand
\name{lolify_in_environment}
\alias{lolify_in_environment}
\alias{unlolify_in_environment}
\title{Lolify Functions in other packages.}
\usage{
lolify_in_environment(sym, env, which = "all")

unlolify_in_environment(sym, env, which = "all")
}
\arguments{
\item{sym}{a name object or character string, referring to the
function to modify (not the function it self, as in
\code{\link{lolify}}).}

\item{env}{The environment in which the function lives.  This
might be 'package:foo'}

\item{which}{As with \code{\link{lolify}}, a character vector
indicating the functions to replace.}
}
\description{
Modify a function in another package to replace calls to
\code{cat} (etc) with lolified verisons.
}
\section{Warning}{

This function is pretty evil, as it modifies aready-loaded code.
Aside from the general sillyness of the package, this would
probably keep the package off CRAN as they take a very dim view of
this.
}
\author{
Rich FitzJohn
}

