% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{LolReporter-class}
\alias{LolReporter}
\alias{LolReporter-class}
\alias{lolreporter}
\title{A Colourful testthat Reporter}
\usage{
lolreporter(...)
}
\arguments{
\item{...}{Arguments passed through to \code{SummaryReporter}}
}
\description{
This can be passed to test_dir and friends to produced rainbow
coloured outputs.  It formats the output very simlarly, but
slightly differently to \code{SummaryReporter}, which is the
default for things like \code{test_dir}.  The differences are: (1)
it breaks the string of dots close to the edge of the screen
(based on \code{getOption("width")} and (2) it produces a summary
saying how many successful and unsuccessful tests were performed.
}

