% Generated by roxygen2 (4.0.1): do not edit by hand
\name{paint}
\alias{paint}
\title{Paint String}
\usage{
paint(string, ...)
}
\arguments{
\item{string}{vector of strings to be coloured}

\item{...}{Options.  Currently valid options are either a single
three-element vector or a three row matrix, both interpreted as
red/green/blue.  R's standard recycling rules are used.}
}
\description{
"Paint" a string by wrapping it in ANSI colour code information.
}
\author{
Rich FitzJohn
}

