% Generated by roxygen2 (4.0.1): do not edit by hand
\name{unpaint}
\alias{unpaint}
\title{Remove Formatting From a String}
\usage{
unpaint(string)
}
\arguments{
\item{string}{A string that was generated by \code{\link{paint}}.}
}
\description{
Remove formatting from a string
}
\author{
Rich FitzJohn
}

