% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lolcat.R
\name{lolcat}
\alias{lolcat}
\alias{lolmessage}
\alias{lolprint}
\title{Taste The Rainbow}
\usage{
lolcat(..., file = "", fill = FALSE, lol = NULL)

lolmessage(..., domain = NULL, appendLF = TRUE, lol = NULL)

lolprint(x, ..., lol = NULL)
}
\arguments{
\item{...}{Arguments as for \link{cat}, including
\code{sep}, etc, arguments passed to \link{message} or extra arguments
to \link{print}}

\item{file}{Used to check that we don't put silly output into
actual files.  Base \link{cat} will be used when \code{file} is
given}

\item{fill}{Passed to \link{cat}}

\item{lol}{Optional \link{lol} engine to use. If omitted
we use the default one registered when the package is loaded}

\item{domain}{used by translations (not yet supported)}

\item{appendLF}{logical: should messages given as a character
string have a newline appended?  (see \link{message})}

\item{x}{Object to print}
}
\description{
Replacement for cat, message, and print.
}
\examples{
for (i in 1:20) {
  lolcat("hello world\n")
}
}
