% Generated by roxygen2 (4.0.1): do not edit by hand
\name{lolcat}
\alias{lolcat}
\alias{lolmessage}
\alias{lolstop}
\alias{lolwarning}
\title{Taste The Rainbow}
\usage{
lolcat(..., file = "")

lolmessage(..., domain = NULL, appendLF = TRUE)

lolwarning(..., call. = TRUE, immediate. = FALSE, noBreaks. = FALSE,
  domain = NULL)

lolstop(..., call. = TRUE, domain = NULL)
}
\arguments{
\item{...}{Arguments as for \code{\link{cat}}, including
\code{sep}, etc.}

\item{file}{Used to check that we don't put silly output into
actual files.  Base \code{cat} will be used when \code{file} is
given.}

\item{domain}{used by translations (not yet supported)}

\item{appendLF}{logical: should messages given as a character
string have a newline appended?  (see \code{\link{message}}).}

\item{call.,immediate.,noBreaks.}{Arguments passed through to
\code{\link{warning}} and \code{\link{stop}} - these are passed
directly through to the underlying function.}
}
\description{
Replacement for cat, message, warning and stop.
}
\details{
Note that the warning and stop verions don't work very well at the
moment, and message does not yet bold the output.
}
\examples{
for (i in 1:20) {
  lolcat("hello world\\n")
}
}
\author{
Rich FitzJohn
}

