% Generated by roxygen2 (4.0.2): do not edit by hand
\name{lolify}
\alias{lolify}
\alias{unlolify}
\title{Make Functions More Colourful}
\usage{
lolify(f, which = "all")

unlolify(f)
}
\arguments{
\item{f}{A function to modify}

\item{which}{Character vector indicating which types of output
generating functions to modify.  Possible values are \code{"cat"},
\code{"message"}, \code{"warning"} and \code{"stop"}, indicating
which functions to modify.  Multiple values are OK (e.g,
\code{c("cat", "message")}.  The special value \code{"all"}
changes all values and is the default.}
}
\value{
A function with the environent rewritten to use lolifed
versions of output functions
}
\description{
x1Convert a function to use rainbow-writing functions: replacing
\code{cat} with \code{lolcat}, etc.  This is done by rewriting
environments rather than rewriting code.
}
\details{
The function \code{unlolify} reverses the process.
}
\author{
Rich FitzJohn
}

